Refactored code in password change.
